# specs/001-georectification/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ALPROJ GUI Backend API
  version: 1.0.0
  description: |
    景観写真オルソ化アプリのバックエンド API。
    フロントエンド（Svelte）から呼び出される。
servers:
  - url: http://localhost:{port}
    description: ローカルサイドカーサーバー
    variables:
      port:
        default: "8765"
...
paths:
...
  /api/georectify/simulate:
    post:
      tags: [georectify]
      summary: シミュレーション画像生成
      operationId: generateSimulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationRequest"
      responses:
        "200":
          description: シミュレーション画像
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResponse"
  /api/georectify/process:
    post:
      tags: [georectify]
      summary: オルソ化処理開始
      operationId: startProcessing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessRequest"
      responses:
        "202":
          description: 処理開始（ジョブ作成）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /api/georectify/export:
    post:
      tags: [georectify]
      summary: GeoTIFF エクスポート
      operationId: exportGeoTiff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "200":
          description: エクスポート完了
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string

  # === Jobs ===
...
  /api/jobs/{jobId}/ws:
    get:
      tags: [jobs]
      summary: ジョブ進捗 WebSocket
      description: |
        WebSocket 接続で進捗を受信。
        メッセージ形式: {"progress": 0.5, "step": "matching", "message": "..."}
      operationId: jobWebSocket
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "101":
          description: WebSocket 接続
components:
  schemas:
...
    CameraParamsValues:
      type: object
      required: [x, y, z, fov, pan, tilt, roll]
      properties:
        x:
          type: number
        y:
          type: number
        z:
          type: number
        fov:
          type: number
          minimum: 1
          maximum: 180
        pan:
          type: number
          minimum: 0
          maximum: 360
        tilt:
          type: number
          minimum: -90
          maximum: 90
        roll:
          type: number
          minimum: -180
          maximum: 180
        a1:
          type: number
          default: 1
        a2:
          type: number
          default: 1
        k1:
          type: number
          default: 0
        k2:
          type: number
          default: 0
        k3:
          type: number
          default: 0
        k4:
          type: number
          default: 0
        k5:
          type: number
          default: 0
        k6:
          type: number
          default: 0
        p1:
          type: number
          default: 0
        p2:
          type: number
          default: 0
        s1:
          type: number
          default: 0
        s2:
          type: number
          default: 0
        s3:
          type: number
          default: 0
        s4:
          type: number
          default: 0
        cx:
          type: number
        cy:
          type: number

    # === Process Result ===
    ProcessResult:
      type: object
      properties:
        gcps:
          type: array
          items:
            $ref: "#/components/schemas/GCP"
        metrics:
          $ref: "#/components/schemas/ProcessMetrics"
        geotiff_path:
          type: string
        log:
          type: array
          items:
            type: string
    GCP:
      type: object
      required: [id, image_x, image_y, geo_x, geo_y, geo_z, enabled]
      properties:
        id:
          type: integer
        image_x:
          type: number
        image_y:
          type: number
        geo_x:
          type: number
        geo_y:
          type: number
        geo_z:
          type: number
        residual:
          type: number
        enabled:
          type: boolean
          default: true
    ProcessMetrics:
      type: object
      required: [rmse, gcp_count, gcp_total]
      properties:
        rmse:
          type: number
        gcp_count:
          type: integer
        gcp_total:
          type: integer
        residual_mean:
          type: number
        residual_std:
          type: number
        residual_max:
          type: number

    # === Simulation ===
    SimulationRequest:
      type: object
      required: [dsm_path, ortho_path, target_image_path, camera_params]
      properties:
        dsm_path:
          type: string
        ortho_path:
          type: string
        target_image_path:
          type: string
        camera_params:
          $ref: "#/components/schemas/CameraParamsValues"
        max_size:
          type: integer
          default: 800
          description: 出力画像の最大サイズ
    SimulationResponse:
      type: object
      required: [image_base64]
      properties:
        image_base64:
          type: string
          description: Base64 エンコードされた PNG 画像

    # === Process ===
    ProcessRequest:
      type: object
      required: [project_id]
      properties:
        project_id:
          type: string
          format: uuid
        options:
          $ref: "#/components/schemas/ProcessOptions"
    ProcessOptions:
      type: object
      properties:
        matching_method:
          type: string
          enum: [akaze, sift, roma, loftr]
          default: akaze
        optimizer:
          type: string
          enum: [cma, lsq]
          default: cma
        max_generations:
          type: integer
          default: 300
        min_gcp_distance:
          type: number
          default: 100
          description: カメラからの最小 GCP 距離 (m)

    # === Export ===
    ExportRequest:
      type: object
      required: [project_id, output_path]
      properties:
        project_id:
          type: string
          format: uuid
        output_path:
          type: string
        resolution:
          type: number
          default: 1.0
          description: 出力解像度 (m/pixel)
        crs:
          type: string
          default: "EPSG:6690"
        interpolate:
          type: boolean
          default: true

    # === Job ===
    Job:
      type: object
      required: [id, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        progress:
          type: number
          minimum: 0
          maximum: 1
        step:
          type: string
          description: 現在のステップ名
        message:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        error:
          type: string
        result:
          $ref: "#/components/schemas/ProcessResult"
...